+++ lib/authoreyes/helpers/in_model.rb
@@ -7,6 +7,8 @@ module Authoreyes
       include InstanceVariables
 
       module ClassMethods
+        # Iterates through records in a collection and calls
+        # include_privileges_for on each.
         def include_privileges_for(user=nil)
           all.map do |item|
             item.include_privileges_for(user)