+++ CHANGELOG.md
@@ -2,6 +2,7 @@
   * Update README to display more comprehensive documentation and quick-start
   * Add `privileges_on` and `guest_privileges_on` helpers
   * Add `InUser` helper module with `is_authoreyes_user` method to simplify setup
+  * Add `include_privileges_for` helper to InModel helpers
 
 * __v0.2.2__
   * Fix some bugs preventing controllers named certain ways not to behave properly